+++ R/mutate_nse.R
@@ -6,9 +6,8 @@
 #'
 #' Mutate a data frame by the mutate terms from \code{...}.
 #'
-#' Note: this method as the default setting \code{mutate_nse_split_terms = TRUE}, which while
-#' safer (avoiding certain known \code{dplyr}/\code{dblyr} issues) can be needlessly expensive
-#' and have its own "too long sequence" issues on remote-data systems
+#' Note: this method as the default setting \code{mutate_nse_split_terms = TRUE}, which is
+#' safer (avoiding certain known \code{dplyr}/\code{dblyr} issues)
 #' (please see the side-notes of \url{http://winvector.github.io/FluidData/partition_mutate.html} for some references).
 #'
 #' @seealso \code{\link{mutate_se}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_at}}, \code{\link[wrapr]{:=}}